requirements.txt
streamlit
apify-client
openai
langchain
python-dotenv
langchain_community